Forcing the webView to reload when going into background in the scan page
Going into background will cause the websocket server to be shut down, in which case a new one must be opened to serve for future connections.
If the user however was in the scan page before going into background, on foreground the image will freeze because the javascript ws will be connected to the websocket server that has just been shut down.

Reloading the page has no effect due to some iframe errors, so as a temporary solution the app will be reloaded from the origin in this case.